# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API EduApp",
        "description": "Documentation Swagger de l'API d'authentification",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/loginn": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Connexion utilisateur",
                "description": "Permet de se connecter et de récupérer un token Bearer",
                "operationId": "388c6bc6e7b0e699059f20b21a7a017d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "test@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Connexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "token_abc123"
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        },
                                        "user": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Identifiants invalides"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentification"
                ],
                "summary": "Déconnexion utilisateur",
                "description": "Supprime le token actuel et déconnecte l'utilisateur",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "description": "Déconnexion réussie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Déconnecté avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Token invalide ou manquant"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/enseignant/{code_enseignant}/groupes": {
            "get": {
                "tags": [
                    "Enseignants"
                ],
                "summary": "Récupérer les groupes d'un enseignant pour une date donnée",
                "description": "Retourne les groupes d'un enseignant pour la date spécifiée en query param",
                "operationId": "1271215fc4278c7de019807a15877dd6",
                "parameters": [
                    {
                        "name": "code_enseignant",
                        "in": "path",
                        "description": "Code de l'enseignant",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date des séances à récupérer (format YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date",
                            "example": "2026-01-10"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des groupes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun groupe trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/groupe/{code_groupe}/etudiants": {
            "get": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Récupérer les étudiants d'un groupe",
                "description": "Retourne tous les étudiants inscrits dans un groupe donné",
                "operationId": "34ba320fe7da6a1b58fd0db132fbab7b",
                "parameters": [
                    {
                        "name": "code_groupe",
                        "in": "path",
                        "description": "Code du groupe",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "G101"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des étudiants",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucun étudiant trouvé pour ce groupe"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/groupe/{code_groupe}/matieres": {
            "get": {
                "tags": [
                    "Groupes"
                ],
                "summary": "Récupérer les matières d'un groupe",
                "description": "Retourne toutes les matières associées à un groupe",
                "operationId": "226fc97c91e82ef6bf577b980fb3ae58",
                "parameters": [
                    {
                        "name": "code_groupe",
                        "in": "path",
                        "description": "Code du groupe",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "G101"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des matières",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Aucune matière trouvée pour ce groupe"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}